refactor Login into PasswordLoginConnected to follow connected component pattern
