+++ packages/apollos-ui-auth/src/PasswordLogin/authenticate.js
