+++ packages/apollos-ui-auth/src/PasswordLogin/index.js
@@ -1,0 +1,4 @@
+import PasswordLogin from './PasswordLogin';
+
+export PasswordLoginConnected from './PasswordLoginConnected';
+export default PasswordLogin;
